Factor out IMMS#resetCurrentMethodAndClient().
As a preparation to fix Bug 25373872, this CL does a mechanical
code clean-ups in InputMethodManagerService.

As a follow up of I6c3186050592526fc95c5b27f18e2155acff5ebc, this CL
introduces IMMS#resetCurrentMethodAndClient() to make it clear when we
are resetting both IMMS->IME and IMMS->Apps connections.

This is literally mechanical.  No behavior change is intended.

Bug: 25373872
Change-Id: Ia4161854d9da71783fd028fcb79407e4b9463aee