Only create surface when showing window if it destroys it when hiding.
The visibility of a window might be toggled to true even when it doesn't
have a surface, which is a case for windows under the lock screen. We
can't blindly create surfaces in that case, but only do it for the
windows that destroy their surfaces when they are hidden.

Bug: 25879215
Change-Id: I6cf2c6810ce02fba0d2207a56de9924c0270dfeb