Add two new intents
One intent notifies the new and previous default sms app when the user
changes which app is the default sms app. The second intent is sent to the
default sms app when some other white-listed app, such as Android Auto or
the phone, modifies the Telephony db behind the default sms app's back.

Change-Id: Iee1e1c4a4c299846f382bd8247c295cf7aaa040e