Add ability to promote fields.
This adds the ability to promote private fields to public ones to
enable layoutlib to access them. This was first added in change
5676482c92a55d30be13ffba51c5f3988cd5bdf4, but reverted in
f805062e7f510163db64234a27416cdc51ff6c60. This is essentially a revert
of the change again, but without the SimpleMonthView related stuff.

Also, promote the Choreographer's field as is needed to support
animations.

Change-Id: Iee45a465c7a1cadd362c1841385941e6232e4900