docs: Removing info about settings activity for channel apps
Settings activity is no longer supported for channel services, so
we're removing that from the docs (and the sample manifest
snippet). Note that the Setup activity is still required, and
now serves double-duty for both channel setup and channel
settings, so I adjusted the section at the end accordingly.

Added updates based on review suggestions.

Bug: 21205118
Change-Id: Ia394b83cee0cb0e7ae6caf87c9d783eb6b4e04ef