Changed the action executed when "Take bug report" is selected.
By default, it executes the new, enhanced bugreport (service
'bugreportplus'), which is more user-friendly (it shows a system
notification with the progress, allow user to cancel, etc...), but
causes more interference in the system operations.

But on long press, it executes the old, lighter workflow (service
'bugreport'), which is less user-friendly but causes less
interference (and hence should be used in the cases where the device is
slow or unresponsive).

BUG: 26034608
Change-Id: I6759e4c8c2c2a970d0a2bb99ca3471854d01780a