Revert "Camera2: create new streams if surface size has changed"
This change is redundant after the following change is merged:

Camera2: Handle surface size change

When comparing if two OutputConfigurations are equal, also check
the width, height, format, dataspace when the output configurations
were created. With this, CameraDeviceImpl knows that the surface has
changed sizes and will reconfigure the stream.

Bug: 12250682
Change-Id: I4216260a0a08ad87b835ed97e9eaddf1df317fcc
This reverts commit ad5882d3dac2285d561b00aa1de46dbcd932365b.