When the device boots up, netd works more than required
When the device boots up, netd adds rules for applications
which do not have the NETWORK permission to iptables.
Therefore, optimize NetworkPolicyManagerService to not touch
uids that do not have the NETWORK permission.
This modification is similar to Google commit fa0a83cb9.

Cherry picked from AOSP commit to master.

Bug: 27165396
Change-Id: Ic8bb837143b9e349859210654248195d62b73d17