Removed outdated TODO for LocaleList
The suggested TODO is no longer deemed necessary (and is too
expensive to do for all LocaleLists). LocaleLists get (and will get)
passed down to the native level through strings that are the result
of calls to toLanguageTags(), which is pre-computed at construction
time. Users who need to interact with the native level can do their
own caching, such as in commit
54711fdffa59ce4a42ba7d2c274c69258d61e802.

Change-Id: I665f3b28eaf8df309bf1aec96f9d237aeb72865c