Do not add far favorite printers more than once.
The favorite printer list contains a printer multiple times as it might
have been used at different locations. Hence it can be that a printer is
in the favorite printer list multipe times but never considered close.

Hence we go into the second loop and add it multiple times as we never
add it to alreadyAddedPrinter.

Change-Id: I4ca8a6a5a05f43645a0b2c891a55df2ccce05ad4