Init child fragments consistently with activity fragments
Fragments contained within an activity are restored from instance
state in FragmentActivity.onCreate, but fragments contained within
another fragment were previously restored from instance state in
performCreate after onCreate returned. This meant that developers
couldn't consistently rely on being able to control when this happens
with a call to super.onCreate, and calls to findFragmentByTag to
reconnect with a restored child fragment would fail in ways that
succeed when the fragment is directly added to an activity.

Change child fragment instance state restore to happen during
Fragment.onCreate instead of in performCreate to be consistent with
activity behavior. Preserve the old behavior for apps targeting
an SDK version < N.

Change-Id: I9c88d5554da9a32219d64c70ca638f75ecb233ed